requirements.txt
coreapi==2.3.3
gdown==4.4.0
keras==2.8.0
mtcnn==0.1.1
numpy==1.22.3
Pillow==9.1.0
psycopg2==2.9.3
PyQt5==5.15.6
scikit_learn==1.0.2
scipy==1.8.0
tensorflow==2.8.0
